Coverted all the PDF figures to EPS.
I used `pdftops -eps` to convert all the PDF figures to EPS, and changed
the figure calls to use the new figures. This means that `pdflatex` now
generates a document with blank spaces where the figures go, and you need
to now build with `latex` and then use `dvipdf` to convert that the
generated DVI to a PDF (if that's your goal).

I also had weird problems where if I view the resulting PDF in Apple's
Preview app all the figures have their colors inverted. If I view it in
Adobe Reader then all looks fine, so hopefully this will work for the
GPTP folks.